+++ src/app/Priceoffer/offerController.js
@@ -10,7 +10,7 @@ const {emit} = require("nodemon");
 /**
  * API No. 23
  * API Name : 가격제안 생성 API
- * [POST] /app/priceoffer
+ * [POST] /app/priceoffer/:userId
  */
 
 exports.postPriceoffer = async function (req, res) {
@@ -18,23 +18,34 @@ exports.postPriceoffer = async function (req, res) {
     /**
      * Body: userId, price, goodsId
      */
-    const {userId, price, goodsId} = req.body;
+    // jwt - userId, path variable :userId
 
-    // 빈 값 체크
-    if (!userId)
-        return res.send(response(baseResponse.USER_USERID_EMPTY));
-        
-    if (!price)
+    const userIdFromJWT = req.verifiedToken.userId
+
+    const userId = req.params.userId;
+    const {price, goodsId} = req.body;
+
+    if (!userId) return res.send(errResponse(baseResponse.USER_USERID_EMPTY));
+
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    } else {
+        // 빈 값 체크
+        if (!price)
         return res.send(response(baseResponse.GOODS_PRICE_EMPTY));
+
+        if (!goodsId)
+            return res.send(response(baseResponse.GOODS_GOODSID_EMPTY));
+
+        const offerResponse = await offerService.createPriceoffer(
+            userId, price, goodsId
+        )
+
+        return res.send(offerResponse);
+    }
+
+
     
-    if (!goodsId)
-        return res.send(response(baseResponse.GOODS_GOODSID_EMPTY));
-
-    const signUpResponse = await offerService.createPriceoffer(
-        userId, price, goodsId
-    )
-
-    return res.send(signUpResponse);
  }
 
 
@@ -49,12 +60,21 @@ exports.getPriceofferById = async function (req, res) {
     /**
      * Path Variable: userId
      */
+    // jwt - userId, path variable :userId
+
+    const userIdFromJWT = req.verifiedToken.userId
+
     const userId = req.params.userId;
 
     if (!userId) return res.send(errResponse(baseResponse.USER_USERID_EMPTY));
 
-    const priceofferByIdResult = await offerProvider.retrievePriceofferById(userId);
-    return res.send(response(baseResponse.SUCCESS, priceofferByIdResult));
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    } else {
+        const priceofferByIdResult = await offerProvider.retrievePriceofferById(userId);
+        return res.send(response(baseResponse.SUCCESS, priceofferByIdResult));
+    }
+    
     
 };
 
@@ -63,21 +83,21 @@ exports.getPriceofferById = async function (req, res) {
  * API Name : 가격제안 수정 API
  * [PUT] /app/priceoffer/:userId
  */
- exports.putPriceofferById = async function (req, res) {
+//  exports.putPriceofferById = async function (req, res) {
 
-    /**
-     * Path Variable: userId
-     * Body : goodsId, price
-     */
-    const userId = req.params.userId;
-    const {price, goodsId} = req.body;
+//     /**
+//      * Path Variable: userId
+//      * Body : goodsId, price
+//      */
+//     const userId = req.params.userId;
+//     const {price, goodsId} = req.body;
 
-    if(!userId) return res.send(errResponse(baseResponse.USER_USERID_EMPTY));
+//     if(!userId) return res.send(errResponse(baseResponse.USER_USERID_EMPTY));
 
-    const editPriceofferInfo = await offerService.editPriceoffer(userId, price, goodsId)
-    return res.send(editPriceofferInfo);
+//     const editPriceofferInfo = await offerService.editPriceoffer(userId, price, goodsId)
+//     return res.send(editPriceofferInfo);
     
-};
+// };
 
 /**
  * API No. 26
@@ -87,14 +107,24 @@ exports.getPriceofferById = async function (req, res) {
 
 exports.deletePriceofferByID = async function (req, res) {
 
+    // jwt - userId, path variable :userId
+
+    const userIdFromJWT = req.verifiedToken.userId
+
     const userId = req.params.userId;
     const goodsId = req.body.goodsId;
 
     if (!userId) return res.send(errResponse(baseResponse.USER_USERID_EMPTY));
 
-    if (!goodsId) return res.send(errResponse(baseResponse.GOODS_GOODSID_EMPTY));
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    } else {
+        if (!goodsId) return res.send(errResponse(baseResponse.GOODS_GOODSID_EMPTY));
 
-    const PriceofferById = await offerService.deletePriceoffer(userId, goodsId);
-    return res.send(response(baseResponse.SUCCESS, PriceofferById));
+        const PriceofferById = await offerService.deletePriceoffer(userId, goodsId);
+        return res.send(response(baseResponse.SUCCESS, PriceofferById));
+    }
+    
+    
  }
